# tools/infer-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP management API. All write operations are forwarded to the C++ engine via Unix domain socket.",
        "title": "infer-server Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/models": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "List loaded models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/infer-server.ModelInfo"
                            }
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Load a model",
                "parameters": [
                    {
                        "description": "Model definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infer-server.ModelLoad"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/infer-server.OkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/models/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Unload a model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infer-server.OkResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "model is in use by a running task",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pipelines": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipelines"
                ],
                "summary": "List pipelines",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/infer-server.PipelineInfo"
                            }
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipelines"
                ],
                "summary": "Add a pipeline",
                "parameters": [
                    {
                        "description": "Pipeline definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infer-server.PipelineCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/infer-server.OkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pipelines/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipelines"
                ],
                "summary": "Remove a pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pipeline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infer-server.OkResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "pipeline is in use by a task",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sources": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "List sources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/infer-server.SourceInfo"
                            }
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Add a source",
                "parameters": [
                    {
                        "description": "Source definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infer-server.SourceCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/infer-server.OkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sources/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Remove a source",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infer-server.OkResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "source is in use by a task",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "List tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/infer-server.TaskInfo"
                            }
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a task binding a source to a pipeline. Starts in stopped state; use POST /tasks/{id}/start to activate.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Add a task",
                "parameters": [
                    {
                        "description": "Task definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infer-server.TaskCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/infer-server.OkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Remove a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infer-server.OkResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tasks/{id}/{action}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Start or stop a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "start",
                            "stop"
                        ],
                        "type": "string",
                        "description": "Action",
                        "name": "action",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infer-server.OkResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infer-server.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "engine unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "infer-server.EdgeConfig": {
            "type": "object",
            "required": [
                "from",
                "to"
            ],
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "infer-server.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "in_use"
                },
                "message": {
                    "type": "string",
                    "example": "source in use by a task: cam0"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "infer-server.InputShape": {
            "type": "object",
            "properties": {
                "c": {
                    "type": "integer",
                    "example": 3
                },
                "h": {
                    "type": "integer",
                    "example": 640
                },
                "w": {
                    "type": "integer",
                    "example": 640
                }
            }
        },
        "infer-server.ModelInfo": {
            "type": "object",
            "properties": {
                "backend": {
                    "type": "string"
                },
                "batch_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "input_shape": {
                    "type": "string",
                    "example": "3x640x640"
                },
                "instance_count": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "infer-server.ModelLoad": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "backend": {
                    "type": "string",
                    "enum": [
                        "cpu",
                        "trt",
                        "ascend"
                    ]
                },
                "batch_size": {
                    "type": "integer"
                },
                "engine_path": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_shape": {
                    "$ref": "#/definitions/infer-server.InputShape"
                },
                "onnx_path": {
                    "type": "string"
                }
            }
        },
        "infer-server.OkResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "cam0"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "infer-server.PipelineCreate": {
            "type": "object",
            "required": [
                "id",
                "nodes"
            ],
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/infer-server.EdgeConfig"
                    }
                },
                "id": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/infer-server.StageConfig"
                    }
                }
            }
        },
        "infer-server.PipelineInfo": {
            "type": "object",
            "properties": {
                "edge_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "infer-server.SourceCreate": {
            "type": "object",
            "required": [
                "id",
                "url"
            ],
            "properties": {
                "degraded_threshold": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "max_reconnect_attempts": {
                    "type": "integer"
                },
                "max_reconnect_delay_ms": {
                    "type": "integer"
                },
                "reconnect_delay_ms": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "infer-server.SourceInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "reconnect_count": {
                    "type": "integer"
                },
                "state": {
                    "type": "string",
                    "enum": [
                        "CONNECTING",
                        "STREAMING",
                        "RECONNECTING",
                        "DEGRADED",
                        "STOPPED"
                    ]
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "infer-server.StageConfig": {
            "type": "object",
            "required": [
                "id",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "with": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "infer-server.TaskCreate": {
            "type": "object",
            "required": [
                "id",
                "pipeline_id",
                "source_id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "sample_fps": {
                    "type": "number"
                },
                "source_id": {
                    "type": "string"
                }
            }
        },
        "infer-server.TaskInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "state": {
                    "type": "string",
                    "enum": [
                        "running",
                        "stopped"
                    ]
                }
            }
        }
    }
}