# tools/infer-server/main.go
// Package main is the infer-server management sidecar.
//
//	@title			infer-server Management API
//	@version		1.0
//	@description	HTTP management API. All write operations are forwarded to the C++ engine via Unix domain socket.
//	@host			localhost:8080
//	@BasePath		/
package main

import (
	"bufio"
	"encoding/json"
	"fmt"
	"net"
	"net/http"
	"os"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	_ "infer-tools/infer-server/docs" // generated by swag init
)

var (
	socketPath = envOr("INFER_SOCKET", "/var/run/infer.sock")
	listenAddr = envOr("INFER_SERVER_ADDR", ":8080")
)

func envOr(key, def string) string {
	if v := os.Getenv(key); v != "" {
		return v
	}
	return def
}

func call(cmd map[string]any) (map[string]any, error) {
	conn, err := net.Dial("unix", socketPath)
	if err != nil {
		return nil, err
	}
	defer conn.Close()

	data, err := json.Marshal(cmd)
	if err != nil {
		return nil, fmt.Errorf("marshal request: %w", err)
	}
	fmt.Fprintf(conn, "%s\n", data)

	var resp map[string]any
	if err := json.NewDecoder(bufio.NewReader(conn)).Decode(&resp); err != nil {
		return nil, err
	}
	return resp, nil
}

func engineUnavailable(c *gin.Context, err error) {
	c.String(http.StatusServiceUnavailable, err.Error())
}

// handleHealthz godoc
//
//	@Summary	Health check
//	@Tags		health
//	@Produce	plain
//	@Success	200	{string}	string	"OK"
//	@Failure	503	{string}	string	"engine unreachable"
//	@Router		/healthz [get]
func handleHealthz(c *gin.Context) {
	resp, err := call(map[string]any{"cmd": "health"})
	if err != nil {
		engineUnavailable(c, err)
		return
	}
	if resp["status"] != "ok" {
		c.JSON(http.StatusServiceUnavailable, resp)
		return
	}
	c.String(http.StatusOK, "OK")
}

// handleMetrics godoc
//
//	@Summary	Prometheus metrics
//	@Tags		health
//	@Produce	plain
//	@Success	200	{string}	string	"Prometheus text format"
//	@Failure	503	{string}	string	"engine unreachable"
//	@Router		/metrics [get]
func handleMetrics(c *gin.Context) {
	resp, err := call(map[string]any{"cmd": "metrics"})
	if err != nil {
		engineUnavailable(c, err)
		return
	}
	c.Header("Content-Type", "text/plain; version=0.0.4; charset=utf-8")
	if m, ok := resp["data"].(string); ok {
		c.String(http.StatusOK, m)
	}
}

// handleListSources godoc
//
//	@Summary	List sources
//	@Tags		sources
//	@Produce	json
//	@Success	200	{array}		SourceInfo
//	@Failure	503	{string}	string	"engine unreachable"
//	@Router		/sources [get]
func handleListSources(c *gin.Context) {
	resp, err := call(map[string]any{"cmd": "list_sources"})
	if err != nil {
		engineUnavailable(c, err)
		return
	}
	c.JSON(http.StatusOK, resp["data"])
}

// handleAddSource godoc
//
//	@Summary	Add a source
//	@Tags		sources
//	@Accept		json
//	@Produce	json
//	@Param		body	body		SourceCreate	true	"Source definition"
//	@Success	201		{object}	OkResponse
//	@Failure	400		{object}	ErrorResponse
//	@Failure	503		{string}	string	"engine unreachable"
//	@Router		/sources [post]
func handleAddSource(c *gin.Context) {
	callCreate(c, "add_source")
}

// handleRemoveSource godoc
//
//	@Summary	Remove a source
//	@Tags		sources
//	@Produce	json
//	@Param		id	path		string	true	"Source ID"
//	@Success	200	{object}	OkResponse
//	@Failure	404	{object}	ErrorResponse
//	@Failure	409	{object}	ErrorResponse	"source is in use by a task"
//	@Failure	503	{string}	string			"engine unreachable"
//	@Router		/sources/{id} [delete]
func handleRemoveSource(c *gin.Context) {
	callDelete(c, "remove_source")
}

// handleListPipelines godoc
//
//	@Summary	List pipelines
//	@Tags		pipelines
//	@Produce	json
//	@Success	200	{array}		PipelineInfo
//	@Failure	503	{string}	string	"engine unreachable"
//	@Router		/pipelines [get]
func handleListPipelines(c *gin.Context) {
	resp, err := call(map[string]any{"cmd": "list_pipelines"})
	if err != nil {
		engineUnavailable(c, err)
		return
	}
	c.JSON(http.StatusOK, resp["data"])
}

// handleAddPipeline godoc
//
//	@Summary	Add a pipeline
//	@Tags		pipelines
//	@Accept		json
//	@Produce	json
//	@Param		body	body		PipelineCreate	true	"Pipeline definition"
//	@Success	201		{object}	OkResponse
//	@Failure	400		{object}	ErrorResponse
//	@Failure	503		{string}	string	"engine unreachable"
//	@Router		/pipelines [post]
func handleAddPipeline(c *gin.Context) {
	callCreate(c, "add_pipeline")
}

// handleRemovePipeline godoc
//
//	@Summary	Remove a pipeline
//	@Tags		pipelines
//	@Produce	json
//	@Param		id	path		string	true	"Pipeline ID"
//	@Success	200	{object}	OkResponse
//	@Failure	404	{object}	ErrorResponse
//	@Failure	409	{object}	ErrorResponse	"pipeline is in use by a task"
//	@Failure	503	{string}	string			"engine unreachable"
//	@Router		/pipelines/{id} [delete]
func handleRemovePipeline(c *gin.Context) {
	callDelete(c, "remove_pipeline")
}

// handleListModels godoc
//
//	@Summary	List loaded models
//	@Tags		models
//	@Produce	json
//	@Success	200	{array}		ModelInfo
//	@Failure	503	{string}	string	"engine unreachable"
//	@Router		/models [get]
func handleListModels(c *gin.Context) {
	resp, err := call(map[string]any{"cmd": "list_models"})
	if err != nil {
		engineUnavailable(c, err)
		return
	}
	c.JSON(http.StatusOK, resp["data"])
}

// handleLoadModel godoc
//
//	@Summary	Load a model
//	@Tags		models
//	@Accept		json
//	@Produce	json
//	@Param		body	body		ModelLoad	true	"Model definition"
//	@Success	201		{object}	OkResponse
//	@Failure	400		{object}	ErrorResponse
//	@Failure	503		{string}	string	"engine unreachable"
//	@Router		/models [post]
func handleLoadModel(c *gin.Context) {
	callCreate(c, "load_model")
}

// handleUnloadModel godoc
//
//	@Summary	Unload a model
//	@Tags		models
//	@Produce	json
//	@Param		id	path		string	true	"Model ID"
//	@Success	200	{object}	OkResponse
//	@Failure	404	{object}	ErrorResponse
//	@Failure	409	{object}	ErrorResponse	"model is in use by a running task"
//	@Failure	503	{string}	string			"engine unreachable"
//	@Router		/models/{id} [delete]
func handleUnloadModel(c *gin.Context) {
	callDelete(c, "unload_model")
}

// handleListTasks godoc
//
//	@Summary	List tasks
//	@Tags		tasks
//	@Produce	json
//	@Success	200	{array}		TaskInfo
//	@Failure	503	{string}	string	"engine unreachable"
//	@Router		/tasks [get]
func handleListTasks(c *gin.Context) {
	resp, err := call(map[string]any{"cmd": "list_tasks"})
	if err != nil {
		engineUnavailable(c, err)
		return
	}
	c.JSON(http.StatusOK, resp["data"])
}

// handleAddTask godoc
//
//	@Summary		Add a task
//	@Description	Creates a task binding a source to a pipeline. Starts in stopped state; use POST /tasks/{id}/start to activate.
//	@Tags			tasks
//	@Accept			json
//	@Produce		json
//	@Param			body	body		TaskCreate	true	"Task definition"
//	@Success		201		{object}	OkResponse
//	@Failure		400		{object}	ErrorResponse
//	@Failure		503		{string}	string	"engine unreachable"
//	@Router			/tasks [post]
func handleAddTask(c *gin.Context) {
	callCreate(c, "add_task")
}

// handleRemoveTask godoc
//
//	@Summary	Remove a task
//	@Tags		tasks
//	@Produce	json
//	@Param		id	path		string	true	"Task ID"
//	@Success	200	{object}	OkResponse
//	@Failure	404	{object}	ErrorResponse
//	@Failure	503	{string}	string	"engine unreachable"
//	@Router		/tasks/{id} [delete]
func handleRemoveTask(c *gin.Context) {
	callDelete(c, "remove_task")
}

// handleTaskAction godoc
//
//	@Summary	Start or stop a task
//	@Tags		tasks
//	@Produce	json
//	@Param		id		path		string	true	"Task ID"
//	@Param		action	path		string	true	"Action"	Enums(start, stop)
//	@Success	200		{object}	OkResponse
//	@Failure	404		{object}	ErrorResponse
//	@Failure	503		{string}	string	"engine unreachable"
//	@Router		/tasks/{id}/{action} [post]
func handleTaskAction(c *gin.Context) {
	id := c.Param("id")
	action := c.Param("action")
	if action != "start" && action != "stop" {
		c.Status(http.StatusNotFound)
		return
	}
	resp, err := call(map[string]any{"cmd": action + "_task", "id": id})
	if err != nil {
		engineUnavailable(c, err)
		return
	}
	code := http.StatusOK
	if resp["status"] != "ok" {
		code = http.StatusNotFound
	}
	c.JSON(code, resp)
}

func callCreate(c *gin.Context, cmd string) {
	var body map[string]any
	if err := c.ShouldBindJSON(&body); err != nil {
		c.String(http.StatusBadRequest, "invalid JSON body")
		return
	}
	body["cmd"] = cmd
	resp, err := call(body)
	if err != nil {
		engineUnavailable(c, err)
		return
	}
	code := http.StatusCreated
	if resp["status"] != "ok" {
		code = http.StatusBadRequest
	}
	c.JSON(code, resp)
}

func callDelete(c *gin.Context, cmd string) {
	id := c.Param("id")
	resp, err := call(map[string]any{"cmd": cmd, "id": id})
	if err != nil {
		engineUnavailable(c, err)
		return
	}
	code := http.StatusOK
	if resp["status"] != "ok" {
		if resp["code"] == "in_use" {
			code = http.StatusConflict
		} else {
			code = http.StatusNotFound
		}
	}
	c.JSON(code, resp)
}

func main() {
	r := gin.Default()

	r.GET("/healthz", handleHealthz)
	r.GET("/metrics", handleMetrics)

	r.GET("/sources", handleListSources)
	r.POST("/sources", handleAddSource)
	r.DELETE("/sources/:id", handleRemoveSource)

	r.GET("/pipelines", handleListPipelines)
	r.POST("/pipelines", handleAddPipeline)
	r.DELETE("/pipelines/:id", handleRemovePipeline)

	r.GET("/models", handleListModels)
	r.POST("/models", handleLoadModel)
	r.DELETE("/models/:id", handleUnloadModel)

	r.GET("/tasks", handleListTasks)
	r.POST("/tasks", handleAddTask)
	r.DELETE("/tasks/:id", handleRemoveTask)
	r.POST("/tasks/:id/:action", handleTaskAction)

	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	fmt.Printf("infer-server listening on %s (socket: %s)\n", listenAddr, socketPath)
	if err := r.Run(listenAddr); err != nil {
		fmt.Fprintln(os.Stderr, err)
		os.Exit(1)
	}
}